Move Matchable into a separate file
